+++ email_auth/interfaces/rest/urls.py
@@ -16,4 +16,9 @@ urlpatterns = [
         views.EmailVerificationView.as_view(),
         name="email-verification-create",
     ),
+    path(
+        "password-reset-requests/",
+        views.PasswordResetRequestView.as_view(),
+        name="password-reset-request-list",
+    ),
 ]